[internal][report] Fix bug in token parsing, add token JSON
`internal/report.parseUploadToken` used to use `fmt.Sscanf`. It turns
out that this function succeeds even if it's only able to parse part
of its input, which means that a string with a prefix which is a
valid upload token was accepted. Fix this by replacing `fmt.Sscanf`
with `srcconv.ParseUint`.