Make handlers simpler, factoring out setup and error handling
Change the signature of handlers so that they take a `*util.Context`
and return a `util.StatusError`. The `main` package is responsible
for providing a handler with the original signature - taking both an
`http.ResponseWriter` and an `*http.Request` - and is responsible
for constructing a `*util.Context` and handling any errors. Once we
have more handlers, this will save us a lot of boilerplate, and will
also reduce the possibility for mistakes (e.g., omitting error
handling in certain paths).